Pixar layout normalization.
This PR introduce a <Column> wrapper and some utility CSS classes for solid and fluid column items.
UI widgets will be anotated with these classes to decide if they should spread over the column left/right padding or not.
By default these classes are noop and it's safe to have them all the time on the widgets. It's only when the widgets are wrspped inside a <Column> element that the new behaviors are activated. Currently
the <Column> is not used (beside some story pages) and will be used to migrate page by page or section by section the new behavior.

I created new pages storybook pages (SRP, messaging, ridedetails) that uses this layout. Ultimetaly all pages could be in the storybook too.

Note that some widgets are currently broken when the new behavior is activated (the Itinirerary for instance). This will be fixed as follow task after the initial commit.

TESTED=Storybook that widgets are not broken when not wrapped in a <Column>