[BBC-11430] Handle `defaultValue` update in `SelectField`
With OneAction & our kairos architecture, we can be sure that the defaultValue will be correctly set a first render (race condition). Since defaultValue can't be updated after it's first rendering, we should use the `value`. We don't want to control this value from outside the component, so I just did as what we implemented in `TextArea` and `TextField`