Use transient props in Address component for styled-comp props
The Address component is currently throwing a lot of warnings due to the highlighted props wich is only used by styled-component but is not meant to be rendered in the DOM. To remove these warnings this PR uses the [transient props syntax of Styled-component][1].

```
Warning: Received `false` for a non-boolean attribute `highlighted`.

If you want to write it to the DOM, pass a string instead: highlighted="false" or highlighted={value.toString()}.

If you used to conditionally omit it with highlighted={condition && value}, pass highlighted={condition ? value : undefined} instead.
```

[1]: https://styled-components.com/docs/api#transient-props